Removing remaining automatic function calling
This changes `{{foo}}` and `{{#foo}}...{{/foo}}` to be Lookup
expressions that do not call `foo` implicitly. This also removes any
special logic for looking up scope properties vs helper functions since
can-view-scope will now check helpers if a property is not found on the
scope.

Lastly, this removes explicit `metadata` from helpers and stores the
`isHelper` and `isLiveBound` properties directly on the helper
functions.