Fix false-positive key lookup warning when key is dotted
When given a key like `foo.bar`, first check if `foo` is defined on the
context, and then check if `bar` is defined on `foo`. Each step of the
way, we only care that `hasKey` is truthy, if the context doesn't exist
at any point, we don't warn, because the lookup will rerun once a value
is set. Fixes #462