fixing tests using Symbols for IE11
IE11 does not support Symbols, so can-symbol falls back to using string
keys. These keys are included when running deepEqual on the result of
expression.parse, whereas they are not when Symbols are used. To fix
these tests, I included the value stored to these keys in the expected
value that is passed to deepEqual.