Portals
This adds support for portals as described in #646

Portals are a way to attach HTML somewhere other than within the
template. This is commonly used for injecting modals into the
document.body, or for adding metadata such as the `document.head`.

In the implementation portals are always *appended* to the host element.
This means they don't destroy any DOM already there. When a template is
removed the DOM portaled to the other element will be removed.

Closes #646.