Add the done.keepNode symbol to portaled nodes
done-autorender, as part of reattachment, removes all nodes within the
`<head>` and `<body>` and replaces it with what was rendered from the
template.

This is a problem for portals which renders outside of the template. The
solution here is to add a symbol to these nodes, which done-autorender
will look for in order to know to leave them alone.